requirements.txt
streamlit
pandas
numpy
lime
google-genai 
python-dotenv
pickle
matplotlib